requirements.txt
# Core dependencies - using compatible versions
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Free LLM options
transformers>=4.20.0
torch>=1.12.0
accelerate>=0.20.0
ollama>=0.1.0
huggingface-hub>=0.16.0

# PDF processing - using compatible versions
PyPDF2>=3.0.1
pytesseract>=0.3.10
# Pillow will be installed automatically with sentence-transformers
pdf2image>=1.16.3

# Web framework
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.1

# Data processing
pandas>=2.1.3
numpy>=1.24.3
pydantic>=2.0.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# LLM Integration
openai>=1.0.0
anthropic>=0.3.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
tqdm>=4.66.1
PyYAML>=6.0
